01-Ai_agents_first/PROJECTS/calculator_agent/pyproject.toml
[project]
name = "calculator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
]
